# docs/openapi.yaml
openapi: 3.0.0
info:
  title: inf-22-dw2-g21
  version: 0.0.1
  description: 'Guarde suas mídias favoritas'
servers:
- url: http://localhost:3000
  description: My localhost
  variables:
    basePath:
      default: /
components:
  schemas:
    User:
      title: User
      type: object
      required:
      - apiKey
      properties:
        id:
          type: integer
          format: int64
        apiKey:
          type: string
        username:
          type: string
        email:
          type: string
      additionalProperties: false
    Filme:
      title: Filme
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        ano:
          type: string
        genero:
          type: string
      required:
        - nome
        - ano
      additionalProperties: false
    Serie:
      title: Serie
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        ano:
          type: string
        genero:
          type: string
      required:
        - nome
        - ano
      additionalProperties: false
...
paths:
  /filmes:
    get:
      tags:
        - FilmesController
      operationId: retrieveFilmes
      summary: Retrieve filmes
      responses:
        '200':
          description: Array of Filmes model instances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Filme'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Filme'
                xml:
                  name: Filmes
    post:
      tags:
        - FilmesController
      operationId: createFilme
      summary: Create filme
      responses:
        '200':
          description: Creat a Filme model instance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Filme'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
          application/xml:
            schema:
              $ref: '#/components/schemas/Filme'
...
  /series:
    get:
      tags:
        - SeriesController
      operationId: retrieveSeries
      summary: Retrieve series
      responses:
        '200':
          description: Array of Series model instances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Serie'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Serie'
                xml:
                  name: Series
    post:
      tags:
        - SeriesController
      operationId: createSerie
      summary: Create serie
      responses:
        '200':
          description: Creat a Serie model instance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Serie'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Serie'
          application/xml:
            schema:
              $ref: '#/components/schemas/Serie'
...
  /users/:
    get:                 
      tags:               
        - User            
      summary: Access to unsecured resource
      security: []  # No security
      produces:
      - text/plain
      responses:          
        200:
          description: Success Message
    post: 
      tags: 
        - User 
      summary: Access to unsecured resource
      operationId: createUser
      security: [] # No security
      produces:
        - text/plain
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
          application/xml:
            schema:
              $ref: '#/components/schemas/User'
          required: true
      responses:
        "200":
          description: Create a User model instance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          description: Bad request
        "404":
          description: id not found
      x-swagger-router-controller: User
...